Add files via upload
CORDIC Processor - Computing the Values of Sine and Cosine - This can also be effectively extrapolated to linear combinations and other complex functions using modules to multiplication and addition. Please note that this code as such is non hardware synthesisable as there is no output - The user needs to edit the code a bit -  give some outputs and then is expected to use an FPGA for printing out the corresponding values.